# rad/api/oas/oas_response_encoders_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"net/http"

	"github.com/go-faster/errors"
	"github.com/go-faster/jx"
	ht "github.com/ogen-go/ogen/http"
	"go.opentelemetry.io/otel/codes"
	"go.opentelemetry.io/otel/trace"
)

func encodeColumnCreateResponse(response ColumnCreateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *TableInfo:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *ColumnCreateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *ColumnCreateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeColumnDeleteResponse(response ColumnDeleteRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *TableInfo:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *ColumnDeleteConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *ColumnDeleteUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeColumnUpdateResponse(response ColumnUpdateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *TableInfo:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *ColumnUpdateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *ColumnUpdateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeGetHealthResponse(response *Health, w http.ResponseWriter, span trace.Span) error {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(200)

	e := new(jx.Encoder)
	response.Encode(e)
	if _, err := e.WriteTo(w); err != nil {
		return errors.Wrap(err, "write")
	}

	return nil
}

func encodeIndexCreateResponse(response IndexCreateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *TableInfo:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *IndexCreateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *IndexCreateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeIndexDeleteResponse(response IndexDeleteRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *TableInfo:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *IndexDeleteConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *IndexDeleteUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeQueryResponse(response QueryRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *QueryResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *QueryBadRequest:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(400)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *QueryUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeRowCreateResponse(response RowCreateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *RecordResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *RowCreateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *RowCreateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeRowDeleteResponse(response RowDeleteRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *DeleteResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *RowDeleteConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *RowDeleteUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeRowUpdateResponse(response RowUpdateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *RecordResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *RowUpdateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *RowUpdateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeSchemaMigrateResponse(response SchemaMigrateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *MigrateResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *Problem:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTableCreateResponse(response TableCreateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *TableInfo:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TableCreateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TableCreateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTableDeleteResponse(response TableDeleteRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *NoContent:
		w.WriteHeader(204)

		return nil

	case *TableDeleteConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TableDeleteUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTableListResponse(response *TableList, w http.ResponseWriter, span trace.Span) error {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(200)

	e := new(jx.Encoder)
	response.Encode(e)
	if _, err := e.WriteTo(w); err != nil {
		return errors.Wrap(err, "write")
	}

	return nil
}

func encodeTableUpdateResponse(response TableUpdateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *TableInfo:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TableUpdateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TableUpdateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTransactionBeginResponse(response *TransactionCredentials, w http.ResponseWriter, span trace.Span) error {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(200)

	e := new(jx.Encoder)
	response.Encode(e)
	if _, err := e.WriteTo(w); err != nil {
		return errors.Wrap(err, "write")
	}

	return nil
}

func encodeTransactionCommitResponse(response TransactionCommitRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *NoContent:
		w.WriteHeader(204)

		return nil

	case *TransactionCommitNotFound:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(404)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionCommitConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTransactionQueryResponse(response TransactionQueryRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *QueryResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionQueryBadRequest:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(400)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionQueryNotFound:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(404)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionQueryUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTransactionRollbackResponse(response TransactionRollbackRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *NoContent:
		w.WriteHeader(204)

		return nil

	case *Problem:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(404)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTransactionRowCreateResponse(response TransactionRowCreateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *RecordResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowCreateNotFound:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(404)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowCreateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowCreateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTransactionRowDeleteResponse(response TransactionRowDeleteRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *DeleteResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowDeleteNotFound:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(404)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowDeleteConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowDeleteUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeTransactionRowUpdateResponse(response TransactionRowUpdateRes, w http.ResponseWriter, span trace.Span) error {
	switch response := response.(type) {
	case *RecordResult:
		w.Header().Set("Content-Type", "application/json; charset=utf-8")
		w.WriteHeader(200)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowUpdateNotFound:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(404)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowUpdateConflict:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(409)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	case *TransactionRowUpdateUnprocessableEntity:
		w.Header().Set("Content-Type", "application/problem+json")
		w.WriteHeader(422)

		e := new(jx.Encoder)
		response.Encode(e)
		if _, err := e.WriteTo(w); err != nil {
			return errors.Wrap(err, "write")
		}

		return nil

	default:
		return errors.Errorf("unexpected response type: %T", response)
	}
}

func encodeErrorResponse(response *InternalServerErrorStatusCode, w http.ResponseWriter, span trace.Span) error {
	w.Header().Set("Content-Type", "application/problem+json")
	code := response.StatusCode
	if code == 0 {
		// Set default status code.
		code = http.StatusOK
	}
	w.WriteHeader(code)
	if code >= http.StatusInternalServerError {
		span.SetStatus(codes.Error, http.StatusText(code))
	}

	e := new(jx.Encoder)
	response.Response.Encode(e)
	if _, err := e.WriteTo(w); err != nil {
		return errors.Wrap(err, "write")
	}

	if code >= http.StatusInternalServerError {
		return errors.Wrapf(ht.ErrInternalServerErrorResponse, "code: %d, message: %s", code, http.StatusText(code))
	}
	return nil

}