# rad/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    name: Barnaby Keene
  description: >
    The Rad wire protocol: a small JSON over HTTP surface that clients and a
    Rad server speak to each other.


    Rad is reached over a `rad://` (or `rads://` for HTTPS) connection URI
    which resolves to an HTTP base URL. Rad speaks plain HTTP; front it with
    your own proxy for encryption or authentication and use `rads://` so the
    client dials it over HTTPS. Rad never terminates TLS itself. The default
    port is 7237.


    This document is the contract. Every operation describes what it does, what
    it returns, and how it fails, in enough detail to serve as the basis for
    acceptance tests. The protocol is deliberately unversioned: Rad is a proof
    of concept and makes no backwards compatibility guarantees yet.


    Errors are returned as RFC 7807 Problem Details with the
    `application/problem+json` media type. The `code` member is the stable,
    machine readable discriminator (`invalid`, `not_found`, `conflict`,
    `internal`) and clients should switch on it rather than parsing the `type`
    URI or the HTTP status.
  title: rad
  version: "v0"
servers:
  - url: "/"
    description: >
      The server mounts every operation at the root. A `rad://host:7237`
      connection URI maps directly onto `http://host:7237/`.
...
paths:
...
  /tables:
    get:
      operationId: TableList
      summary: List the tables in the database.
      description: |
        Return every table currently defined in the catalog, along with its
        columns and primary key. This reflects the schema as it exists right
        now, after the most recent successful migration. An empty database
        returns an empty list, not an error.

        Use this to introspect a running server without access to the original
        `schema.rad` file.
      tags: [meta]
      responses:
        default: { $ref: "#/components/responses/InternalServerError" }
        "200": { $ref: "#/components/responses/TableListOK" }
    post:
      operationId: TableCreate
      summary: Create a table.
      description: |
        Define a new table in one call: columns, primary key, and optionally
        indexes and foreign keys, exactly as a `schema.rad` entry would. IDs
        are assigned by the catalog and the whole definition commits
        atomically — a rejected definition leaves nothing behind, including
        the name.

        Foreign keys may reference existing tables or the table being created
        (self-references), and must target the referenced table's full
        primary key. A definition that fails validation — duplicate or
        missing name, unsupported column type, nullable primary key column,
        index or key over unknown columns — is rejected with an `invalid`
        problem. On a schema-managed database this operation is always
        rejected.
      tags: [catalog]
      requestBody: { $ref: "#/components/requestBodies/TableCreate" }
      responses:
        default: { $ref: "#/components/responses/InternalServerError" }
        "200": { $ref: "#/components/responses/TableOK" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
components:
...
  requestBodies:
...
    TableCreate:
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TableDef" }
...
  responses:
    InternalServerError:
      description: Something went wrong inside the server that was not the caller's fault.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Conflict:
      description: The write lost an optimistic race or violated a constraint.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnprocessableEntity:
      description: The request was well formed but failed validation against the catalog.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    TableListOK:
      description: The current set of tables.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TableList" }
    TableOK:
      description: The table's definition after the mutation was applied.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TableInfo" }
...
  schemas:
    Problem:
      description: |
        An RFC 7807 Problem Details object. Every non 2xx response carries one.
        The `code` member mirrors the last segment of `type` so clients can
        branch on a stable string without parsing the URI.
      type: object
      required: [type, title, status, code]
      properties:
        type:
          description: A stable URN identifying the problem kind, for example `urn:rad:problem:conflict`.
          type: string
        title:
          description: A short, human readable summary of the problem kind.
          type: string
        status:
          description: The HTTP status code, repeated in the body for convenience.
          type: integer
        detail:
          description: A human readable explanation specific to this occurrence.
          type: string
        code:
          description: |
            The machine readable discriminator. One of `invalid` for a malformed
            or rejected request, `execution_failed` for a valid query that
            failed on the data it met (division by zero, a violated cardinality
            assertion), `not_found`, `conflict` for a lost optimistic race or
            constraint violation, or `internal`.
          type: string
          enum: [invalid, execution_failed, not_found, conflict, internal]
    Value:
      description: An arbitrary JSON value carried by the HTTP protocol.
...
    ColumnDefault:
      description: |
        A column default, applied when an insert omits the column: either a
        builtin generator named by `func` (`uuid` on text columns, `now_ms`
        on int64 columns) or a literal `value` of the column's type. Exactly
        one of the two is set.
      type: object
      properties:
        func:
          description: A builtin generator, `uuid` or `now_ms`.
          type: string
        value:
          description: A literal of the column's type.
          allOf:
            - $ref: "#/components/schemas/Value"
    ColumnInfo:
      description: One column of a table, as reported by introspection.
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          description: The column's storage type, one of `text`, `int64`, `float64`, or `bool`.
          type: string
        nullable:
          type: boolean
        format:
          description: An optional semantic hint such as `uuid` or `unix_ms`.
          type: string
        default:
          $ref: "#/components/schemas/ColumnDefault"
    IndexInfo:
      description: One secondary index, in both definitions and introspection.
      type: object
      required: [name, columns]
      properties:
        name:
          type: string
        columns:
          description: The indexed column names, in order.
          type: array
          items:
            type: string
        unique:
          type: boolean
    ForeignKeyInfo:
      description: |
        One foreign key, in both definitions and introspection. The
        referenced columns must be the referenced table's full primary key.
      type: object
      required: [name, columns, ref_table, ref_columns]
      properties:
        name:
          type: string
        columns:
          description: The referencing column names on this table.
          type: array
          items:
            type: string
        ref_table:
          description: The referenced table's name.
          type: string
        ref_columns:
          description: The referenced table's primary key columns, in order.
          type: array
          items:
            type: string
    TableInfo:
      description: One table's definition, as reported by introspection.
      type: object
      required: [name, columns, primary_key]
      properties:
        name:
          type: string
        columns:
          type: array
          items: { $ref: "#/components/schemas/ColumnInfo" }
        primary_key:
          description: The column names that make up the primary key, in order.
          type: array
          items:
            type: string
        indexes:
          type: array
          items: { $ref: "#/components/schemas/IndexInfo" }
        foreign_keys:
          type: array
          items: { $ref: "#/components/schemas/ForeignKeyInfo" }
    TableDef:
      description: |
        A new table's definition: exactly what a `schema.rad` entry
        expresses, as JSON. Column types are `text`, `int64`, `float64`, or
        `bool`; the primary key is required and its columns must not be
        nullable.
      type: object
      required: [name, columns, primary_key]
      properties:
        name:
          type: string
        columns:
          type: array
          items: { $ref: "#/components/schemas/ColumnInfo" }
        primary_key:
          type: array
          items:
            type: string
        indexes:
          type: array
          items: { $ref: "#/components/schemas/IndexInfo" }
        foreign_keys:
          type: array
          items: { $ref: "#/components/schemas/ForeignKeyInfo" }
...
    TableList:
      description: The set of tables defined in the database.
      type: object
      required: [tables]
      properties:
        tables:
          type: array
          items: { $ref: "#/components/schemas/TableInfo" }
...